+++ modules/video/include/opencv2/video/detail/lapjv.hpp
@@ -1,3 +1,4 @@
+#include <stdlib.h>
 #ifndef LAPJV_H
 #define LAPJV_H
