+++ modules/video/src/tracking/detail/byteTracker_strack.cpp
@@ -2,15 +2,16 @@
 // It is subject to the license terms in the LICENSE file found in the top-level directory
 // of this distribution and at http://opencv.org/license.html.
 
-#include "../../precomp.hpp"
-#include "bytetracker_strack.hpp"
+// #include "bytetracker_strack.hpp"
+//#include "opencv2/video/detail/bytetracker_strack.hpp"
+#include "opencv2/video/tracking.hpp"
 //#include "opencv2/core.hpp"
 
 namespace cv {
-namespace detail {
-namespace tracking {
+//namespace detail {
+//namespace tracking {
     
-using namespace cv::detail::tracking;
+//using namespace cv::detail::tracking;
 
 Strack::Strack()
 {
@@ -146,7 +147,7 @@ float Strack::getScore() const
 }
 
 }
-}
-}
+//}
+//}
